requirements.txt
autopep8==1.4.4
entrypoints==0.3
flake8==3.7.8
mccabe==0.6.1
pycodestyle==2.5.0
pycryptodome==3.9.0
pyflakes==2.1.1
